Finish function A.

+++ src/main/java/sima/antwaffle/types/Ant.java
@@ -4,11 +4,15 @@ package sima.antwaffle.types;
  * Common attributes for ants.
  */
 public abstract class Ant {
-    public final long id;
+
+    /**
+     * Zero indexed.
+     */
+    public final int id;
     public final double x;
     public final double y;
 
-    protected Ant(long id, double x, double y) {
+    protected Ant(int id, double x, double y) {
         this.id = id;
         this.x = x;
         this.y = y;
